Migrate suite_test.go to grafana folder
I need to have one suite_test.go per controller. But at the same time if I have multiple
I get port issues and probably other issues as well.
It will also be hard to test multiple things at the same time, which we need to do with
grafana and grafanadashboards for example